refactor: extract runtime launch config
