+++ crates/agent-engine/src/runtime/child_agents/launch_context.rs
@@ -1,5 +1,5 @@
 use crate::runtime::agent_loop::RuntimeLoopState;
-use crate::runtime::engine::AgentConfig;
+use crate::runtime::launch_config::AgentConfig;
 use alan_agent_protocol::{GovernanceConfig, SpawnHandle, SpawnSpec, SpawnTarget};
 use anyhow::{Context, Result, bail};
 use std::path::PathBuf;
